+++ jest.config.js
@@ -1,3 +1,0 @@
-export default {
-  transformIgnorePatterns: ['<rootDir>/node_modules/(?!lodash-es)'],
-};